requirements.txt
playwright>=1.49.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
python-dotenv>=1.0.0
